Refactor parser
Closes #1
