+++ src/finder/mod.rs
@@ -1,2 +1,3 @@
+pub mod location;
 pub mod lottery;
-pub mod lottery_location;
+pub mod parser;
